Refactor Navbar component structure and logic
- Split the Navbar into multiple components: `NavbarBrandSection`, `NavbarMenuItems`, `NavbarUserMenu`, and `NavbarMobileMenu` for better organization and maintainability.
- Introduced a custom hook `useNavbarLogic` to manage the state and behavior of the Navbar, including menu open state and active item tracking.
- Moved the `getMenuIcon` function to a separate file for improved modularity.
- Updated the site configuration to reflect the new project name and description.
